requirements.txt
Flask

# Application server for both development and production.
gunicorn

#
flask-bootstrap

# Testing and static analysis.
pytest
pytest-cov
flake8

# CLI.
Click

# Data and workers.
psycopg2
Flask-SQLAlchemy
redis
celery

# Forms.
Flask-WTF
WTForms-Components

# Extensions.
flask-debugtoolbar
Flask-Mail
flask-moment

#Nov 22
Werkzeug
itsdangerous
Flask-Login